extract copyable to own component
